+++ Namukkun/src/main/java/com/pard/namukkun/login/dto/KakaoUserInfoResponseDto.java
@@ -179,7 +179,6 @@ public class KakaoUserInfoResponseDto {
 
         }
     }
-
     @Getter
     @NoArgsConstructor
     @JsonIgnoreProperties(ignoreUnknown = true)
