--- Namukkun/src/main/java/com/pard/namukkun/login/dto/LoginCreatUserDTO.java
@@ -1,11 +1,0 @@
-package com.pard.namukkun.login.dto;
-
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginCreatUserDTO {
-    private Integer local;
-}
