+++ Namukkun/src/main/java/com/pard/namukkun/login/service/KakaoService.java
@@ -37,11 +37,7 @@ public class KakaoService {
     }
 
     public String getAccessTokenFromKakao(String code) {
-
-//        log.info(clientSecretId);
         try {
-
-
             KakaoTokenResponseDto kakaoTokenResponseDto = WebClient.create(KAUTH_TOKEN_URL_HOST).post()
                     .uri(uriBuilder -> uriBuilder
                             .scheme("https")
@@ -59,8 +55,6 @@ public class KakaoService {
                     .onStatus(HttpStatusCode::is5xxServerError, clientResponse -> Mono.error(new RuntimeException("Internal Server Error")))
                     .bodyToMono(KakaoTokenResponseDto.class)
                     .block();
-
-
             System.out.println(" [Kakao Service] Access Token ------> " + kakaoTokenResponseDto.getAccessToken());
             System.out.println(" [Kakao Service] Refresh Token ------> " + kakaoTokenResponseDto.getRefreshToken());
             //제공 조건: OpenID Connect가 활성화 된 앱의 토큰 발급 요청인 경우 또는 scope에 openid를 포함한 추가 항목 동의 받기 요청을 거친 토큰 발급 요청인 경우
@@ -100,7 +94,6 @@ public class KakaoService {
 
             return userInfo;
         } catch (Exception e) {
-            System.out.println("!!" + e.getMessage());
             return null;
         }
     }
